# web-api/docs/swagger.yaml
basePath: '{{.BasePath}}'
definitions:
  model.Account:
    properties:
      address:
        type: string
      c_time:
        type: string
      change_time:
        type: string
      contact:
        type: string
      email:
        type: string
      id:
        type: integer
      ll_time:
        type: string
      nick_name:
        type: string
      phone:
        type: string
      pid:
        type: integer
      privilege_id:
        type: integer
      pwd:
        type: string
      remark:
        type: string
      role_id:
        type: integer
      state:
        type: string
      username:
        example: elephant
        type: string
    type: object
  model.AccountDeviceTransReq:
    properties:
      devices:
        items:
          $ref: '#/definitions/model.Device'
        type: array
      receiver:
        $ref: '#/definitions/model.DeviceReceiver'
        type: object
      sender:
        $ref: '#/definitions/model.Account'
        type: object
    type: object
  model.AccountImportDeviceReq:
    properties:
      d_type:
        type: string
      devices:
        items:
          $ref: '#/definitions/model.Device'
        type: array
    type: object
...
  model.Device:
    properties:
      account_id:
        type: integer
      active_time:
        type: string
      change_time:
        type: string
      course:
        type: number
      create_time:
        type: string
      device_type:
        type: string
      gps_data:
        $ref: '#/definitions/model.GPS'
        type: object
      id:
        type: integer
      imei:
        type: string
      last_login_time:
        type: string
      local_time:
        type: integer
      nick_name:
        type: string
      password:
        type: string
      sale_time:
        type: string
      speed:
        type: number
      user_name:
        type: string
    type: object
  model.DeviceReceiver:
    properties:
      account_id:
        type: integer
      account_name:
        type: string
    type: object
  model.DeviceUpdate:
    properties:
      ActiveTime:
        type: string
      ChangeTime:
        type: string
      CreateTime:
        type: string
      DeviceType:
        type: string
      IMei:
        type: string
      Id:
        type: integer
      LLTime:
        type: string
      LocalTime:
        type: integer
      LoginId:
        type: integer
      NickName:
        type: string
      SaleTime:
        type: string
    type: object
  model.GPS:
    properties:
      lat:
        type: number
      lng:
        type: number
    type: object
host: '{{.Host}}'
info:
  contact: {}
  description: jimi co ptt platform web api.
  license: {}
  title: JiMi Web API
  version: 1.0.1
paths:
...
  /account/login.do/{account_name}:
    post:
      consumes:
      - application/json
      description: login by account name and pwd
      parameters:
      - description: elephant
        in: path
        name: account_name
        required: true
        type: string
      - description: 登录信息，username和pwd必填
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.Account'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":"true","session_id": sId }'
          schema:
            type: string
      summary: 登录
  /account/logout.do/{account_name}:
    post:
      consumes:
      - application/json
      description: logout by account name and pwd, 请求头中Authorization参数设置为登录时返回的sessionId
      parameters:
      - description: 登录时返回的sessionId
        in: header
        name: Authorization
        required: true
        type: string
      - description: 登录信息，username和pwd必填
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.Account'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":"true","msg": "SignOut is successful"}'
          schema:
            type: string
      summary: 退出
...
  /account_device/{accountId}:
    post:
      parameters:
      - description: 登录时返回的sessionId
        in: header
        name: Authorization
        required: true
        type: string
      - description: 当前用户的账号Id
        in: path
        name: accountId
        required: true
        type: string
      - description: 转移设备的model
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/model.AccountDeviceTransReq'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: '{"result": "success","msg":"Password changed successfully"}'
          schema:
            type: string
      summary: 转移设备
...
  /device/import/{account_name}:
    post:
      parameters:
      - description: 登录时返回的sessionId
        in: header
        name: Authorization
        required: true
        type: string
      - description: 当前用户的账号
        in: path
        name: account_name
        required: true
        type: string
      - description: 导入设备的model
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/model.AccountImportDeviceReq'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: '{"result": "success","msg":"Password changed successfully"}'
          schema:
            type: string
      summary: 导入设备
  /device/update:
    post:
      parameters:
      - description: 登录时返回的sessionId
        in: header
        name: Authorization
        required: true
        type: string
      - description: 当前用户的账号
        in: path
        name: account_name
        required: true
        type: string
      - description: 更新设备信息的model
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/model.DeviceUpdate'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: '{"result": "success","msg":"Password changed successfully"}'
          schema:
            type: string
      summary: 更新设备信息
...
swagger: "2.0"